Dramatically speed up writes
Use a sketchy in-memory buffer to cache writes until they have all arrived, then flush them in a single HTTP call.

This was driven by not being able to do partial PUT, it's simply not supported: https://www.rfc-editor.org/rfc/rfc9110.html#name-partial-put

So I had to find some other way to get the chunks to the Solid server. I had to have all of the chunks arranged before I could send them as it has to be a single PUT.

Although I could use PATCH, servers such as PodSpaces only have `Accept-Patch=application/sparql-update` so I can't patch the contents of a non-RDF Resource which is important because my use case of writing Portable Network Graphics files is the one that highlighted how bad `write()` was behaving with the recently-changed caching strategy meaning the ETag was always changing causing a full re-read every 128KiB.

I genuinely expected a lot more problems with multi-threading on this.

The buffer is bad in many ways. For example it never times out if the flush never gets here. It's very poorly managed with other methods. For example, should read use the buffer. The fuselib architecture has `open()` and `release()` to manage this better but I'm not using that yet.