Create a lot more variety in the tests
Use parameterization to create a lot of different files and folders. This current ends up being 820 tests.

For example we'll stress creating folders with different awkward characters, such as this one:

/test/SolidFS/2024-04-25T10_33_37.362640_00_00/test_file_in_deep_subfolder_ - -\U0001f996_/ / /🦖/

Note that the `_ - -\U0001f996_` part is due to pytest using safe names for the paramterized tests and we're using that to give each test a unique place to work in.